Change extension of fortran sources to F90
The default extension for fortran sources used by autoconf (for
configure tests) is ".f" . Changing this default extension to
".F90".

All fortran sources in this library already use the ".F90"
extension.

This change allows mpi-serial to be built with other libraries
that force fortran sources to have the ".F90" extension (For
example, ACME/CESM/CIME build systems specify compiler flags
that force all fortran sources to use the "*.F90" extension).

This change should not affect libraries that use ".f" or other
suffixes for fortran sources from using the mpi-serial library
(this change is internal to the mpi-serial library)