Rearrange code and other minor fixes needed for compilation with clang
I got a number of compilation errors when building mpi-serial with Apple
clang version 13.1.6, using -std=gnu99 (which is what we use for
building CESM). This commit fixes those compilation errors.

Most of the changes here are just rearranging code so that functions are
declared before they are referenced. However, there are two additional
changes:
- removed ierr from FGet_address
- declared Type_extent in type.h (needed because this function is
  referenced in some other files)